Fixes #24 Move to VS2017 (#26)
* Fixes #24 Move to VS2017

* Fixes #24 Move to VS2017

* Fixes #24 Move to VS2017

* Fixes #24 Move to VS2017

* Force appveyor to use VS2017

* Fixes #24 Move to VS2017

* Fixes #24 Move to VS2017

* Activated tests for ReleaseX64 on Appveyor

+++ src/OSPSuite.SimModel/managed/src/Conversions.cpp
@@ -54,7 +54,7 @@ namespace SimModelNET
 
 	array<double>^ CPPToNETConversions::DoubleVectorToArray(const std::vector<double> & doubleVec)
 	{
-		array<double>^ doubleArray = gcnew array<double>(doubleVec.size());
+		array<double>^ doubleArray = gcnew array<double>((unsigned int)doubleVec.size());
 
 		for(unsigned int i=0; i<doubleVec.size(); i++)
 			doubleArray[i] = doubleVec[i];
@@ -77,7 +77,7 @@ namespace SimModelNET
 
 	array<System::String^>^ CPPToNETConversions::StringVectorToArray(const std::vector<std::string> & stringVec)
 	{
-		array<System::String^>^ stringArray = gcnew array<System::String^>(stringVec.size());;
+		array<System::String^>^ stringArray = gcnew array<System::String^>((unsigned int)stringVec.size());;
 
 		for(unsigned int i=0; i<stringVec.size(); i++)
 			stringArray[i] = MarshalString(stringVec[i]);
